Updated so that this works with Windows
I imported platform to check if it work with windows. The original "self.root" command tries to download put the tmp file in the C:/ directory (at the very top) which causes a file not found error. 

If LCC is set to true, the code as it was did not work with windows.